# labassist-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  admin.AdminStatsResponse:
    properties:
      accepted_applications:
        example: 50
        type: integer
      open_courses:
        example: 5
        type: integer
      pending_applications:
        example: 30
        type: integer
      total_applications:
        example: 200
        type: integer
      total_courses:
        example: 15
        type: integer
      total_instructors:
        example: 10
        type: integer
      total_students:
        example: 80
        type: integer
      total_users:
        example: 100
        type: integer
    type: object
  admin.CreateUserRequest:
    properties:
      full_name:
        example: John Doe
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        example: instructor
    required:
    - full_name
    - role
    type: object
  admin.DeleteCoursesByTermResponse:
    properties:
      deleted:
        example: 12
        type: integer
    type: object
...
  admin.ImportCoursesResponse:
    properties:
      created:
        items:
          $ref: '#/definitions/admin.ImportCourseResult'
        type: array
      skipped:
        items:
          $ref: '#/definitions/admin.ImportSkippedRow'
        type: array
    type: object
...
  admin.UpdateUserRequest:
    properties:
      full_name:
        example: John Doe
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        example: instructor
    type: object
  admin.UpdateUserStatusRequest:
    properties:
      is_active:
        example: true
        type: boolean
    type: object
...
  handlers.BulkReviewResponse:
    properties:
      updated:
        example: 3
        type: integer
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handlers.GoogleLoginRequest:
    properties:
      credential:
        example: eyJhbGci...
        type: string
    required:
    - credential
    type: object
  handlers.GoogleLoginResponse:
    properties:
      is_new_user:
        example: false
        type: boolean
      token:
        example: eyJhbGci...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.LoginRequest:
    properties:
      password:
        example: secret
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handlers.LoginResponse:
    properties:
      token:
        example: eyJhbGci...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  handlers.MeResponse:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.MessageResponse:
    properties:
      message:
        example: ออกจากระบบแล้ว
        type: string
    type: object
...
  models.Course:
    properties:
      academic_year:
        type: integer
      applicant_count:
        type: integer
      capacity:
        type: integer
      code:
        type: string
      created_at:
        type: string
      credits:
        type: string
      deadline:
        type: string
      description:
        type: string
      english_title:
        type: string
      enrolled:
        type: integer
      group_note:
        type: string
      has_lab:
        type: boolean
      id:
        type: integer
      instructor_id:
        type: integer
      instructor_name:
        type: string
      instructors_raw:
        description: |-
          InstructorsRaw is the verbatim text of the spreadsheet's instructor
          cell (kept for display as-is, e.g. co-teaching names the system has
          no matching account for) — separate from InstructorID, which only
          ever points to a real, matched user account.
        type: string
      labboy_accepted:
        type: integer
      labboy_slots:
        type: integer
      requirements:
        type: string
      schedule:
        type: string
      section:
        type: integer
      semester:
        type: string
      status:
        $ref: '#/definitions/models.CourseStatus'
      ta_accepted:
        type: integer
      ta_slots:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        description: |-
          Email has no gorm uniqueIndex tag: admin-created accounts start with a
          blank email (filled in later via Google sign-in), so uniqueness is
          enforced instead by a partial index that ignores empty strings — see
          database/connection.go.
        type: string
      faculty:
        type: string
      full_name:
        type: string
      google_sub:
        type: string
      gpa:
        type: number
      id:
        type: integer
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/models.UserRole'
      student_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
      year:
        type: integer
    type: object
...
  teacher.BulkReviewRequest:
    properties:
      application_ids:
        items:
          type: integer
        type: array
      note:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.AppStatus'
        example: accepted
    required:
    - application_ids
    - status
    type: object
...
  teacher.ReviewRequest:
    properties:
      note:
        example: ผ่านการคัดเลือก
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.AppStatus'
        example: accepted
    required:
    - status
    type: object
...
  teacher.UpdateCourseStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/models.CourseStatus'
        example: open
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API สำหรับระบบจัดการ TA ภาควิชา CS มหาวิทยาลัยศิลปากร
  title: LabAssist API
  version: "1.0"
paths:
  /admin/courses/import:
    post:
      consumes:
      - multipart/form-data
      description: แสดงชื่อผู้สอนตามที่พิมพ์มาในไฟล์เสมอ ระบบพยายามจับคู่ชื่อแรกกับผู้ใช้
        role=instructor ที่มีอยู่จริงเพื่อผูกวิชากับบัญชีนั้น (ให้อาจารย์ล็อกอินดูวิชาของตัวเองได้)
        แต่ถ้าจับคู่ไม่ได้ก็ไม่มีผลต่อการนำเข้า รหัสวิชาซ้ำกันได้ (คนละกลุ่มเรียน)
        มีเพียงแถวที่ว่างจริงๆ เท่านั้นที่จะถูกข้าม
      parameters:
      - description: 'ไฟล์ Excel รายวิชา .xlsx (คอลัมน์: รายวิชา, ชื่อรายวิชา, COURSENAME,
          หน่วยกิต, กลุ่ม, รับ, ลงทะเบียนแล้ว, เวลา, ผู้สอน)'
        in: formData
        name: file
        required: true
        type: file
      - description: ภาคการศึกษา
        in: formData
        name: semester
        required: true
        type: string
      - description: ปีการศึกษา
        in: formData
        name: academic_year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ImportCoursesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: นำเข้ารายวิชาจากไฟล์ Excel (.xlsx)
      tags:
      - admin
  /admin/courses/term:
    delete:
      description: ใช้เพื่อล้างข้อมูลที่นำเข้าผิดพลาด เช่น หลังจากอัปโหลดไฟล์ Excel
        ผิดเทอม
      parameters:
      - description: ภาคการศึกษา
        in: query
        name: semester
        required: true
        type: string
      - description: ปีการศึกษา
        in: query
        name: academic_year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.DeleteCoursesByTermResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลบวิชาทั้งหมดของภาคการศึกษาที่ระบุ
      tags:
      - admin
...
  /admin/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminStatsResponse'
      security:
      - BearerAuth: []
      summary: สถิติภาพรวมของระบบ
      tags:
      - admin
  /admin/users:
    get:
      parameters:
      - description: กรองตาม role
        enum:
        - student
        - instructor
        - staff
        - admin
        in: query
        name: role
        type: string
      - description: ค้นหาด้วยชื่อหรืออีเมล
        in: query
        name: search
        type: string
      - description: จำนวนต่อหน้า (default 100)
        in: query
        name: limit
        type: integer
      - description: ออฟเซ็ต (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerAuth: []
      summary: รายการผู้ใช้ทั้งหมด
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลผู้ใช้
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: สร้างผู้ใช้ใหม่ (ชื่อ + บทบาทเท่านั้น)
      tags:
      - admin
  /admin/users/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: ข้อมูลที่ต้องการแก้ไข
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขชื่อ/บทบาทผู้ใช้
      tags:
      - admin
  /admin/users/{id}/courses:
    get:
      description: จับคู่ด้วย ID บัญชีโดยตรง และด้วยชื่อกับข้อมูลผู้สอนที่นำเข้าจากไฟล์
        Excel (ครอบคลุมทั้งผู้สอนหลักและผู้ร่วมสอน)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายวิชาที่อาจารย์คนนี้สอน
      tags:
      - admin
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: สถานะใหม่
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: อัพเดตสถานะผู้ใช้ (เปิด/ปิด)
      tags:
      - admin
  /auth/google:
    post:
      consumes:
      - application/json
      parameters:
      - description: Google credential
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GoogleLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: เข้าสู่ระบบด้วย Google
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลล็อกอิน
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: เข้าสู่ระบบ
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      security:
      - BearerAuth: []
      summary: ออกจากระบบ
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ดูข้อมูลผู้ใช้ปัจจุบัน
      tags:
      - auth
  /courses:
    get:
      parameters:
      - description: กรองตามสถานะ
        enum:
        - open
        - closing_soon
        - closed
        - draft
        - archived
        in: query
        name: status
        type: string
      - description: ค้นหาด้วยชื่อหรือรหัสวิชา
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: รายการวิชาทั้งหมด (สาธารณะ)
      tags:
      - courses
  /courses/{id}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Course'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: ดูรายละเอียดวิชา (สาธารณะ)
      tags:
      - courses
  /instructor/applications/{id}/review:
    put:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: ผลการตรวจสอบ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ตรวจสอบใบสมัคร (อาจารย์/สตาฟ/แอดมิน)
      tags:
      - instructor
  /instructor/applications/bulk-review:
    put:
      consumes:
      - application/json
      parameters:
      - description: รายการใบสมัครและผลการตรวจสอบ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.BulkReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BulkReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ตรวจสอบใบสมัครแบบกลุ่ม
      tags:
      - instructor
  /instructor/course-catalog:
    get:
      description: จับคู่ด้วยชื่ออาจารย์กับข้อมูลผู้สอนที่นำเข้าจากไฟล์ Excel เพื่อจำกัดตัวเลือกรหัสวิชาให้เหลือแค่วิชาที่ตัวเองสอนจริง
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
      security:
      - BearerAuth: []
      summary: รายวิชาที่อาจารย์คนนี้สอน (สำหรับ dropdown ตอนสร้างประกาศ)
      tags:
      - instructor
  /instructor/courses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายการวิชาของอาจารย์
      tags:
      - instructor
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลวิชา
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: สร้างวิชาใหม่
      tags:
      - instructor
  /instructor/courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ลบวิชา
      tags:
      - instructor
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: ข้อมูลที่ต้องการแก้ไข
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Course'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขข้อมูลวิชา
      tags:
      - instructor
  /instructor/courses/{id}/applicants:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: กรองตามประเภทที่สมัคร
        enum:
        - ta
        - labboy
        in: query
        name: role_applied
        type: string
      - description: กรองตามสถานะ
        enum:
        - pending
        - accepted
        - rejected
        - withdrawn
        in: query
        name: status
        type: string
      - description: ค้นหาด้วยชื่อหรือรหัสนักศึกษา
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Application'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายชื่อผู้สมัครของวิชา
      tags:
      - instructor
  /instructor/courses/{id}/notify:
    post:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ส่งแจ้งเตือนนักศึกษาที่ผ่านการคัดเลือก
      tags:
      - instructor
  /instructor/courses/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: สถานะใหม่
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateCourseStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Course'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: อัพเดตสถานะวิชา
      tags:
      - instructor
  /instructor/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: ข้อมูลโปรไฟล์ของอาจารย์ (ตัวเอง)
      tags:
      - instructor
    put:
      consumes:
      - application/json
      description: บันทึกลง database ทันที การเปลี่ยนแปลงจะแสดงในหน้าจัดการผู้ใช้ของแอดมินด้วย
        เพราะดึงจากข้อมูลผู้ใช้ชุดเดียวกัน
      parameters:
      - description: ข้อมูลโปรไฟล์ที่ต้องการแก้ไข
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateInstructorProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: แก้ไขโปรไฟล์ของอาจารย์ (ตัวเอง)
      tags:
      - instructor
  /student/applications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Application'
            type: array
      security:
      - BearerAuth: []
      summary: รายการใบสมัครของนักศึกษา
      tags:
      - student
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลการสมัคร
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ApplyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: สมัครเป็น TA หรือ Lab Boy
      tags:
      - student
  /student/applications/{id}/withdraw:
    put:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ถอนใบสมัคร
      tags:
      - student
  /student/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StudentDashboardResponse'
      security:
      - BearerAuth: []
      summary: หน้าหลักนักศึกษา
      tags:
      - student
  /student/notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
      security:
      - BearerAuth: []
      summary: รายการแจ้งเตือนของนักศึกษา
      tags:
      - student
  /student/notifications/{id}/read:
    put:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: ทำเครื่องหมายแจ้งเตือนว่าอ่านแล้ว
      tags:
      - student
  /student/notifications/read-all:
    put:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: ทำเครื่องหมายแจ้งเตือนทั้งหมดว่าอ่านแล้ว
      tags:
      - student
  /student/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: ดูโปรไฟล์นักศึกษา
      tags:
      - student
    put:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลโปรไฟล์ที่ต้องการแก้ไข
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: แก้ไขโปรไฟล์นักศึกษา
      tags:
      - student
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"